Move eloquent related code out of EsIndexable
