+++ app/Models/Elasticsearch/BeatmapsetTrait.php
@@ -21,12 +21,12 @@
 namespace App\Models\Elasticsearch;
 
 use App\Models\Beatmap;
-use App\Traits\EsIndexable;
+use App\Traits\EsModelIndexable;
 use Carbon\Carbon;
 
 trait BeatmapsetTrait
 {
-    use EsIndexable;
+    use EsModelIndexable;
 
     public function toEsJson()
     {
