+++ app/Models/Elasticsearch/UserTrait.php
@@ -20,13 +20,13 @@
 
 namespace App\Models\Elasticsearch;
 
-use App\Traits\EsIndexable;
+use App\Traits\EsModelIndexable;
 use Carbon\Carbon;
 use Schema;
 
 trait UserTrait
 {
-    use EsIndexable;
+    use EsModelIndexable;
 
     public function toEsJson()
     {
