+++ app/Traits/EsIndexable.php
@@ -27,8 +27,6 @@ trait EsIndexable
 {
     abstract public static function esIndexName();
 
-    abstract public static function esIndexingQuery();
-
     abstract public static function esSchemaFile();
 
     abstract public static function esType();
@@ -48,7 +46,7 @@ trait EsIndexable
 
     public function getEsId()
     {
-        return $this->getKey();
+        // null will result in an id assigned by elastic
     }
 
     public function esDeleteDocument(array $options = [])
@@ -92,17 +90,15 @@ trait EsIndexable
         return Es::getClient()->indices()->create($params);
     }
 
-    public static function esIndexIntoNew($batchSize = 1000, $name = null, callable $progress = null)
+    public static function esIndexIntoNew(array $options = [], callable $progress = null)
     {
-        $newIndex = $name ?? static::esIndexName().'_'.time();
+        $newIndex = $options['index'] ?? static::esIndexName().'_'.time();
         Log::info("Creating new index {$newIndex}");
         static::esCreateIndex($newIndex);
 
-        $options = [
-            'index' => $newIndex,
-        ];
+        $options['index'] = $newIndex;
 
-        static::esReindexAll($batchSize, 0, $options, $progress);
+        static::esReindexAll($options, $progress);
 
         return $newIndex;
     }
@@ -112,55 +108,7 @@ trait EsIndexable
         return static::esSchemaConfig()['mappings'][static::esType()]['properties'];
     }
 
-    public static function esReindexAll($batchSize = 1000, $fromId = 0, array $options = [], callable $progress = null)
-    {
-        $dummy = new static();
-        $isSoftDeleting = method_exists($dummy, 'getDeletedAtColumn');
-        $startTime = time();
-
-        $baseQuery = static::esIndexingQuery()->where($dummy->getKeyName(), '>', $fromId);
-        $count = 0;
-
-        $baseQuery->chunkById($batchSize, function ($models) use ($options, $isSoftDeleting, &$count, $progress) {
-            $actions = [];
-
-            foreach ($models as $model) {
-                $next = $model;
-                // bulk API am speshul.
-                $metadata = [
-                    '_id' => $model->getEsId(),
-                    'routing' => $model->esRouting(),
-                ];
-
-                if ($isSoftDeleting && $model->trashed()) {
-                    $actions[] = ['delete' => $metadata];
-                } else {
-                    // index requires action and metadata followed by data on the next line.
-                    $actions[] = ['index' => $metadata];
-                    $actions[] = $model->toEsJson();
-                }
-            }
-
-            if ($actions !== []) {
-                $result = Es::getClient()->bulk([
-                    'index' => $options['index'] ?? static::esIndexName(),
-                    'type' => static::esType(),
-                    'body' => $actions,
-                    'client' => ['timeout' => 0],
-                ]);
-
-                $count += count($result['items']);
-            }
-
-            Log::info(static::class." next: {$models->last()->getKey()}");
-            if ($progress) {
-                $progress($count);
-            }
-        });
-
-        $duration = time() - $startTime;
-        Log::info(static::class." Indexed {$count} records in {$duration} s.");
-    }
+    abstract public static function esReindexAll(array $options = [], callable $progress = null);
 
     public static function esSchemaConfig()
     {
